+++ src/main/java/com/example/FriendListApplication/FriendListApp/Repository/FriendRepository.java
@@ -1,8 +1,6 @@
 package com.example.FriendListApplication.FriendListApp.Repository;
 
-
 import com.example.FriendListApplication.FriendListApp.Entity.Friend;
 import org.springframework.data.repository.CrudRepository;
-
-public interface FriendRepository extends CrudRepository<Friend, Integer> {
+public interface FriendRepository extends CrudRepository<Friend, Long> {
 }
